requirements.txt
langchain
wikipedia
langchain-community
python-dotenv
pydantic
duckduckgo-search
google-generativeai
